Keyboard control works now
Method confusion to stop moving in Player (an unused method and an empty one - guess which was being called)
Speed in player adjuted - it was basically crossing the entire board every infinitely small time unit
Forgot to implement listening for the spacebar
Limits were added so that the player cannot go off the board or out of the lower 20% of the screen which they are allowed.
It's lots of fun.